+++ test/IntegrationTest.cpp
@@ -12,17 +12,17 @@
 
 TEST_CASE("Integration", "[integration]")
 {
-    glm::dvec3 const force = pegasus::integration::IntegrateForce({0, 0, 0}, {1, 1, 1});
+    glm::vec3 const force = pegasus::integration::IntegrateForce({0, 0, 0}, {1, 1, 1});
     REQUIRE(
-        true == epona::fp::IsZero(glm::distance(force, glm::dvec3{1, 1, 1}))
+        true == epona::fp::IsZero(glm::distance(force, glm::vec3{1, 1, 1}))
     );
 }
 
 TEST_CASE("Angular integration torque", "[integration]")
 {
-    glm::dvec3 const torque{ 0, 1, 0 };
+    glm::vec3 const torque{ 0, 1, 0 };
     pegasus::mechanics::Body body;
-    body.angularMotion.torque = pegasus::integration::IntegrateTorque(glm::dvec3{ 0 }, torque);
+    body.angularMotion.torque = pegasus::integration::IntegrateTorque(glm::vec3{ 0 }, torque);
     pegasus::integration::Integrate(body, 1.0);
 
     REQUIRE(
